% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_peptides.R
\name{annotate_peptides}
\alias{annotate_peptides}
\title{Annotate peptides based on their enzyme specificity and position within protein sequence.}
\usage{
annotate_peptides(
  peptide2protein,
  fasta,
  decoy_tag = "^rev",
  specificity = "R|K"
)
}
\arguments{
\item{peptide2protein}{A data frame with at least two columns: `Peptide` (peptide sequence to annotate) and `Genes` (Uniprot ID of associated protein).}

\item{fasta}{A list of identified protein sequences, at least matching the ones present in the `peptide2protein` data frame. This list should should be loaded into R with the `read.fasta` function from the `seqinr` package. You need to set the argument as.string=TRUE.}

\item{decoy_tag}{A string with the tag of the decoy sequences.}

\item{specificity}{A string defining the enzyme specificity.}
}
\value{
a data frame with annotation information of the input peptides.
}
\description{
Annotate peptides based on their enzyme specificity and position within protein sequence.
}
