% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_nterm.R
\name{categorize_nterm}
\alias{categorize_nterm}
\title{Categorize and count N-terminal peptides based on Uniprot processing annotation}
\usage{
categorize_nterm(annotated_peptides, uniprot_features, distinct = TRUE)
}
\arguments{
\item{annotated_peptides}{data frame of annotated peptides. The output of the `annotate_nterm` function.}

\item{uniprot_features}{a list of data frames. Each data frame containing Uniprot-annotated features of proteins identified with N-terminal peptides, as resulting from the `feature_to_dataframe` function of the `drawProteins` package.}

\item{nterm}{a character vector. Defines the type of modification expected to be found in N-terminal peptides. `c("TMT-labelled", "acetylated")` is default.}
}
\value{
a list with 4 elements:
        `ntermini_n_processing_features`: data frame of N-terminally identified proteins and their annotated processing features.
        `ntermini_category`: data frame of categorized N-terminal peptides according to their matching to annotated processing sites.
        `ntermini_category_counts`: data frame of counts of matched features and locations to annotated processing sites.
        `ntermini_category_plot`: ggplot2 object. Visualization of counts of matched features and locations.
}
\description{
Categorize and count N-terminal peptides based on Uniprot processing annotation
}
