% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_location_nterm.R
\name{count_location_nterm}
\alias{count_location_nterm}
\title{Generete counts of N-terminal TMT-labelling and acetylation per peptide}
\usage{
count_location_nterm(nterm_annot)
}
\arguments{
\item{nterm_annot}{data frame. Output of `annotate_nterm` function.}
}
\value{
a list. containing tabular counts and visualizations.
}
\description{
Generete counts of N-terminal TMT-labelling and acetylation per peptide
}
