% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fasta}
\alias{fasta}
\title{Protein sequences of identified proteins}
\format{
A list 5986 elements, each represents a protein sequence identified in the Polycystic Disease mouse model experiment.
}
\usage{
fasta
}
\description{
Protein sequences of identified proteins in the Polycystic Disease mouse model experiment
}
\keyword{datasets}
