% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_specific_fdr.R
\name{feature_fdr_correction}
\alias{feature_fdr_correction}
\title{Apply multiple-testing p-value adjustment on specified features}
\usage{
feature_fdr_correction(toptable, interesting_features_table, method = "BH")
}
\arguments{
\item{toptable}{a data frame in the same format as the output of the `topTable` function from the `limma` package}

\item{method}{a string to be passed as a method to the `p.adjust` function for multiple-testing correction. "BH" is default.}
}
\value{
a data frame containing a new column `fdr_correction`, indicating if
        the the adjusted p-value comes from
}
\description{
Apply multiple-testing p-value adjustment on specified features
}
