% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cleave_area.R
\name{get_cleave_area}
\alias{get_cleave_area}
\title{Prepare sequences for motif analysis of cleavage areas}
\usage{
get_cleave_area(peptide_annotation)
}
\arguments{
\item{peptide_annotation}{A data frame. Output from the `annotate_peptides` function.}
}
\value{
A data frame containing information of cleavage areas after proteolytic processing for iceLogo analysis.
}
\description{
Prepare sequences for motif analysis of cleavage areas
}
