% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_peptide_counts.R
\name{list_peptides}
\alias{list_peptides}
\title{Title}
\usage{
list_peptides(annotated_peptide_ids)
}
\arguments{
\item{annotated_peptide_ids}{data frame with annotated peptide IDs including specificity and N-terminal modification.}
}
\value{
A list. Each element is a vector of peptide sequences for each interesting feature.
}
\description{
Title
}
