% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{our_volcano}
\alias{our_volcano}
\title{Generate a volcano plot from a `limma::topTable` output}
\usage{
our_volcano(
  dataarg,
  FC_cutoff = 1,
  pval_cutoff = 0.05,
  color_diffex = "red",
  color_nondifex = "#2a9d8f",
  interesting_proteins = NULL,
  vert_line_col = "red",
  hline_col = "red",
  hline_pos = 0.05,
  linetype = "dashed",
  increased_in,
  comparison_title
)
}
\arguments{
\item{dataarg}{data frame. Output from `limma::topTable`. It should also contain at least
a column called `Protein` with protein IDs.}

\item{comparison_title}{}
}
\value{
ggplot2 object
}
\description{
Generate a volcano plot from a `limma::topTable` output
}
