% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pept_abund_mat}
\alias{pept_abund_mat}
\title{Peptide abundance data frame}
\format{
A data frame with 21032 rows and 22 variables:
\describe{
  \item{index}{Protein and peptide sequence}
  \item{gene}{Gene identifier}
  \item{protein_id}{Protein identifier}
  \item{max_pep_prob}{Max Peptide Probability}
  \item{reference_intensity}{Normalized intensity values of virtual normalization channel}
  \item{ko/wt}{Normalized intensity values of protein by sample}
  \item{mt_*}{Normalized intensity values of empty TMT channels}
}
}
\source{

}
\usage{
pept_abund_mat
}
\description{
Normalized values of abundance of identified proteins from the Polycystic Disease mouse model experiment
}
\keyword{datasets}
