% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pept_ident}
\alias{pept_ident}
\title{Peptides identified and associated modifications}
\format{
A data frame with 30879 rows and 33 variables:
\describe{
  \item{peptide}{Peptide sequence}
  \item{prev_aa}{Amino acid before in the protein sequence}
  \item{next_aa}{Amino acid after in the protein sequence}
  \item{peptide_length}
  \item{probability}{Peptide Probability Score}
  \item{spectral_count}{Spectral counts for identified peptide}
  \item{intensity}{Raw intensity value of precursor ion}
  \item{assigned_modifications}{Chemical modifications assigned to the identified peptide}
  \item{observed_modifications}
  \item{protein}{Protein header}
  \item{protein_id}{Uniprot ID}
  \item{entry_name}{Uniprot entry name}
  \item{gene}
  \item{protein_description}
  \item{mapped_genes}
  \item{mapped_proteins}{Number of unique peptides identified}
  \item{ko/wt}{Raw intensity values of reporter ions by sample}
  \item{mt_*}{Raw intensity values of empty TMT channels}
}
}
\usage{
pept_ident
}
\description{
Peptides identification table from the Polycystic Disease mouse model experiment
}
\keyword{datasets}
