% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_peptide_abund_by_protein_abund.R
\name{peptide2protein_normalization}
\alias{peptide2protein_normalization}
\title{Normalize peptide abundance by protein abundance}
\usage{
peptide2protein_normalization(
  peptides,
  annot,
  peptide_annot,
  summarize_by_specificity = TRUE
)
}
\arguments{
\item{peptides}{Data frame of peptide intensities that will be normalized by protein abundance. Typically gotten from the `purify_n_normalize` function.}

\item{annot}{Data frame obtained from loading the `annotation.txt` from Fragpipe output. First column should be named `channel` and second column should be named `sample`.}

\item{peptide_annot}{peptide annotation with specificity information. As it comes from merging the outputs from the functions `annotate_peptides` and `psmtsv_to_modified_peptides`}
}
\value{
List with four elements:
}
\description{
Takes information of protein intensity and peptide intensity and perform normalization of peptide abundance based on protein abundance.
}
