% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide2matrix.R
\name{peptide_matrix_count}
\alias{peptide_matrix_count}
\title{Convert a character vector of peptides to a matrix of amino acids and count the occurrence of each amino acid in each position}
\usage{
peptide_matrix_count(peptides, p_number = 5)
}
\arguments{
\item{peptides}{A character vector of peptides}

\item{p_number}{An integer value. How many elements before and after the cleavage site to show.}
}
\value{
A list containing two elements:
- peptide_matrix: a matrix where each row represents a peptide and each column represents an amino acid position within that peptide.
Each element of the matrix is a single character representing the amino acid at that position.
- amino_acid_count: a matrix where each row represents an amino acid and each column represents an amino acid position within the peptide.
The intersection represents the number of times that amino acid showed up in that position.
}
\description{
This function takes a character vector of peptides, where each element represents a peptide of length 20, and converts it into a matrix where each row represents a peptide and each column represents an amino acid position within that peptide. Each element of the matrix is a single character representing the amino acid at that position. Additionally, the function counts the number of times a particular amino acid appears in a particular position and generates a new matrix, in which each column is an amino acid position, and each row represents an amino acid. The intersection represents the number of amino acids that showed up in that position.
}
\examples{
peptides <- c("AKKGSKRHYVWALDMTFEH", "VAVFDELLKIVPNLGSYKR")
peptide_matrix_count(peptides)

}
