% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prot_abund_mat}
\alias{prot_abund_mat}
\title{Protein abundance data frame}
\format{
A data frame with 5238 rows and 21 variables:
\describe{
  \item{index}{Protein identifier}
  \item{number_psm}{Number of Peptide-to-spectrum matches for the protein}
  \item{max_pep_prob}{Max Peptide Probability}
  \item{reference_intensity}{Normalized intensity values of virtual normalization channel}
  \item{ko/wt}{Normalized intensity values of protein by sample}
  \item{mt_*}{Normalized intensity values of empty TMT channels}
}
}
\source{

}
\usage{
prot_abund_mat
}
\description{
Normalized values of abundance of identified proteins from the Polycystic Disease mouse model experiment
}
\keyword{datasets}
