% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prot_ident}
\alias{prot_ident}
\title{Proteins identified and associated modifications}
\format{
A data frame with 5986 rows and 41 variables:
\describe{
  \item{group}{Protein group}
  \item{sub_group}{Protein subgroup}
  \item{protein}{Protein header}
  \item{protein_id}{Uniprot protein ID}
  \item{entry_name}{Uniprot Entry ID}
  \item{gene}{Gene name}
  \item{length}{Protein length}
  \item{percent_coverage}{Coverage of protein identification}
  \item{organism}
  \item{protein_description}
  \item{protein_existence}
  \item{protein_description}
  \item{protein_probability}
  \item{top_peptide_probability}
  \item{total_peptides}
  \item{unique_peptides}{Number of unique peptides identified}
  \item{razor_peptides}
  \item{total_spectral_count}{Total spectral counts of identified peptides}
  \item{unique_spectral_count}{Spectral counts of unique peptides}
  \item{razor_spectral_count}{Spectral counts of razor peptides}
  \item{total_intensity}
  \item{unique_intensity}
  \item{razor_intensity}
  \item{razor_assigned_modifications}
  \item{razor_observed_modifications}
  \item{indistinguishable_proteins}
  \item{indistinguishable_proteins}
  \item{ko/wt}{Raw intensity values of reporter ions by sample}
  \item{mt_*}{Raw intensity values of empty TMT channels}
}
}
\usage{
prot_ident
}
\description{
Protein identification table from the Polycystic Disease mouse model experiment
}
\keyword{datasets}
