% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_quant_sum_from_best_psms.R
\name{proteins_from_psms}
\alias{proteins_from_psms}
\title{Proteins from unique PSMs}
\usage{
proteins_from_psms(psms, annot)
}
\arguments{
\item{psms}{Data frame. Optimally, the best PSMs obtained after processing the `psm.tsv` with the function `psmtsv_to_modified_peptides`.}

\item{annot}{Data frame obtained from loading the `annotation.txt` from Fragpipe output. First column should be named `channel` and second column should be named `sample`.}
}
\value{
Data frame of summed protein intensities from unique modified peptides.
}
\description{
Generate protein-level abundance matrix from unique peptides from `psm.tsv` file after selection of modified peptides by best PSMs.
}
