% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_modified_peptides_from_psm_file.R
\name{psmtsv_to_modified_peptides}
\alias{psmtsv_to_modified_peptides}
\title{Process `psm.tsv` file into modified peptides}
\usage{
psmtsv_to_modified_peptides(psms, annot, peptide_probability, minimal_purity)
}
\arguments{
\item{psms}{Data frame obtained from loading the `psm.tsv` file from Fragpipe output}

\item{annot}{Data frame obtained from loading the `annotation.txt` from Fragpipe output. First column should be named `channel` and second column should be named `sample`.}

\item{peptide_probability}{Minimal peptide probability accepted per PSM}

\item{minimal_purity}{Minimal isobaric purity accepted per PSM}
}
\value{
Data frame of summarized modified peptides based on their 'best PSMs' based on the filtering criteria
}
\description{
Filters the `psm.tsv` based on a 'best PSM' criteria; based on purity and summed TMT reporter intensities. Zero (0) abundance values are substituted by NA (missing values).
}
