% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purify_n_normalize.R
\name{purify_n_normalize}
\alias{purify_n_normalize}
\title{Purify quant and normalize}
\usage{
purify_n_normalize(psms, annot)
}
\arguments{
\item{psms}{Data frame. Optimally, the best PSMs obtained after processing the `psm.tsv` with the function `psmtsv_to_modified_peptides`.}

\item{annot}{Data frame obtained from loading the `annotation.txt` from Fragpipe output. First column should be named `channel` and second column should be named `sample`.}
}
\value{
Data frame of summed protein intensities from unique modified peptides.
}
\description{
Refine intensity values based on isobaric purity and normalize. Zero values are substituted for missing values.
}
