% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_annotation}
\alias{sample_annotation}
\title{Sample annotation table}
\format{
A data frame with 30879 rows and 33 variables:
\describe{
  \item{channel}{TMT channel}
  \item{sample_id}{Sample ID}
  \item{condition}{Condition}
}
}
\source{

}
\usage{
sample_annotation
}
\description{
Annotation table of samples and conditions for use case experiment of
     polycystic kidney disease in mice.
}
\keyword{datasets}
