% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sample_proteins}
\alias{sample_proteins}
\title{Pseudo-random sampling of protein IDs from a character vector}
\usage{
sample_proteins(x, size, seed = 363)
}
\arguments{
\item{x}{character vector of protein IDs}

\item{size}{numeric. Number of proteins to sample}

\item{seed}{numeric. Random seed number}
}
\value{
character vector of randomly sampled protein IDs.
}
\description{
Pseudo-random sampling of protein IDs from a character vector
}
