% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_peptide_counts.R
\name{summarize_peptide_counts}
\alias{summarize_peptide_counts}
\title{Summarize counts of peptide identifications}
\usage{
summarize_peptide_counts(annotated_peptide_ids)
}
\arguments{
\item{annotated_peptide_ids}{data frame with annotated peptide IDs including specificity and N-terminal modification.}
}
\value{
table in data frame format.
}
\description{
Summarize counts of peptide identifications
}
